% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_response.R
\name{get_api_response}
\alias{get_api_response}
\title{Send API request}
\usage{
get_api_response(base_url, path_append, query_params = NULL, dry_run = FALSE)
}
\arguments{
\item{base_url}{API endpoint}

\item{path_append}{Any additional paths for API call}

\item{query_params}{Any additional query parameters for API call}

\item{dry_run}{If TRUE, call httr2::req_dry_run, which shows the API call without actually sending it.}
}
\value{

}
\description{
Send API request
}
