% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_url.R
\name{get_base_url}
\alias{get_base_url}
\title{Lookup function for REST API base endpoint URL}
\usage{
get_base_url(exchange)
}
\arguments{
\item{exchange}{Which exchange to use}
}
\value{

}
\description{
Lookup function for REST API base endpoint URL
}
