% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_append.R
\name{get_path_append}
\alias{get_path_append}
\title{Lookup function for REST API URL path appends}
\usage{
get_path_append(exchange, fn, base_asset = NULL, quote_asset = NULL)
}
\arguments{
\item{exchange}{Which exchange}

\item{fn}{Which function needs a path append}

\item{base_asset}{Base asset}

\item{quote_asset}{Quote asset}
}
\value{

}
\description{
Lookup function for REST API URL path appends
}
