% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_query_params.R
\name{get_query_params}
\alias{get_query_params}
\title{Lookup function for REST API URL query parameters}
\usage{
get_query_params(
  exchange,
  fn,
  base_asset = NULL,
  quote_asset = NULL,
  ...,
  level = NULL,
  depth = NULL,
  time_frame = NULL,
  start_time = NULL,
  end_time = NULL,
  limit = NULL,
  candle_type = NULL
)
}
\arguments{
\item{exchange}{Which exchange}

\item{fn}{Which function needs a query}

\item{base_asset}{Base asset}

\item{quote_asset}{Quote asset}

\item{...}{Additional parameters to pass to API call}

\item{level}{Parameter used for varying aggregations of order book.}

\item{depth}{Optional parameter used to specify number of bids and asks to return.}

\item{time_frame}{Parameter used for chart interval and timeframe.}

\item{start_time}{Parameter used for chart beginning time.}

\item{end_time}{Parameter used for chart end time.}

\item{limit}{Parameter used to limit number of candles.}

\item{candle_type}{Type of candles.}
}
\value{

}
\description{
Lookup function for REST API URL query parameters
}
