% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_resp.R
\name{get_tidy_resp}
\alias{get_tidy_resp}
\title{Transform API response into tidy format}
\usage{
get_tidy_resp(exchange, fn, base_asset, quote_asset, resp)
}
\arguments{
\item{exchange}{Which exchange to use for price and market data.}

\item{fn}{Which function}

\item{base_asset}{Base asset}

\item{quote_asset}{Quote asset}

\item{resp}{Response object}
}
\value{

}
\description{
Transform API response into tidy format
}
