% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_balance.R
\name{private_balance}
\alias{private_balance}
\title{Get current account balances
Requires API key access}
\usage{
private_balance(exchange, portfolio = NULL, dry_run = FALSE)
}
\arguments{
\item{exchange}{Which exchange to use. Supported value is "coinbase-pro".}

\item{portfolio}{Optional value if you have multiple portfolios with separate API keys in an exchange.
This must match the set API key value in the system environment.
Use \code{\link{Sys.getenv}} to check what variables are saved if you are uncertain.}

\item{dry_run}{If TRUE, call httr2::req_dry_run, which shows the API call without actually sending it.
In this case, there may be some discrepancy since the actual call will use httr to send, not httr2.}
}
\value{

}
\description{
Get current account balances
Requires API key access
}
