% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_asset_info.R
\name{public_asset_info}
\alias{public_asset_info}
\title{GET additional asset information}
\usage{
public_asset_info(
  exchange = "coinbase-pro",
  base_asset = "BTC",
  quote_asset = "USD"
)
}
\arguments{
\item{exchange}{Which exchange to use for price and market data.
Choices are "bittrex", "coinbase-pro", "gemini", and "kucoin".}

\item{base_asset}{Base asset (default "BTC")}

\item{quote_asset}{Quote asset (default "USD")}
}
\value{

}
\description{
GET additional asset information
}
\examples{
public_asset_info("bittrex")
public_asset_info("coinbase-pro")
public_asset_info("gemini")
public_asset_info("kucoin")
}
