% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_order_book.R
\name{public_order_book}
\alias{public_order_book}
\title{GET current public order book data}
\usage{
public_order_book(
  exchange = "binance",
  base_asset = "BTC",
  quote_asset = "USD",
  ...
)
}
\arguments{
\item{exchange}{Which exchange to use for price and market data. Choices are binance, binance-us, bitstamp, coinbase-pro,
crypto.com, ftx, ftx-us, gemini, kraken, and kucoin.}

\item{base_asset}{Base asset (default BTC)}

\item{quote_asset}{Quote asset (default USD)}

\item{...}{Query parameters passed to API call}
}
\value{

}
\description{
GET current public order book data
}
\examples{
public_order_book(exchange = "binance")
}
