% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_ticker_price.R
\name{public_ticker_price}
\alias{public_ticker_price}
\title{GET current public ticker price and market data}
\usage{
public_ticker_price(
  exchange = "binance",
  base_asset = "BTC",
  quote_asset = "USD",
  price_only = TRUE,
  ...
)
}
\arguments{
\item{exchange}{Which exchange to use for price and market data. Choices are binance, binance-us, bitstamp, coinbase, coinbase-pro,
crypto.com, ftx, ftx-us, gemini, huobi, kraken, and kucoin.}

\item{base_asset}{Base asset (default BTC)}

\item{quote_asset}{Quote asset (default USD)}

\item{price_only}{If TRUE, only give current price as a numeric vector. If false, give all market data}

\item{...}{Other parameters passed to API call}
}
\value{

}
\description{
GET current public ticker price and market data
}
\examples{
public_ticker_price("binance")
}
