% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_trades.R
\name{public_trades}
\alias{public_trades}
\title{GET recent trades}
\usage{
public_trades(
  exchange = "binance",
  base_asset = "BTC",
  quote_asset = "USD",
  ...,
  limit = NULL,
  time_frame = NULL,
  start_time = NULL,
  end_time = NULL
)
}
\arguments{
\item{exchange}{Which exchange to use for price and market data. Choices are "binance", "binance-us", "bittrex", bitstamp", "coinbase",
"coinbase-pro", "crypto.com", "ftx", "ftx-us", "gemini", "huobi", "kraken", "kucoin", and "poloniex".}

\item{base_asset}{Base asset (default BTC)}

\item{quote_asset}{Quote asset (default USD)}

\item{...}{Query parameters passed to API call}

\item{limit}{Parameter used to limit number of trades.
\itemize{
 \item Optional for "binance" and "binance-us". Default: 500; Maximum: 1000.
 \item Optional for "coinbase-pro". Default: 1000; Maximum: 1000.
 \item Optional for "gemini". Default: 50; Maximum: 500.
 \item Optional for "huobi". Default: 1; Maximum: 2000.
 }}

\item{time_frame}{Parameter used for trade interval and timeframe.
\itemize{
 \item Optional for "bitstamp". The time interval from which we want the transactions to be returned.
       Default: "day". Possible options are "minute", "hour", or "day".
}}

\item{start_time}{Parameter used for trade beginning time.
\itemize{
 \item Optional for "ftx" and "ftx-us".
 \item Optional for "gemini". Only return trades after this timestamp. If not present, will show the most recent trades.
 \item Optional for "kraken".
 \item Optional for "poloniex". Return up to 1,000 trades between a range specified in UNIX timestamps.
}}

\item{end_time}{Parameter used for trade end time.
\itemize{
 \item Optional for "ftx" and "ftx-us".
 \item Optional for "poloniex". Return up to 1,000 trades between a range specified in UNIX timestamps.
}}
}
\value{

}
\description{
GET recent trades
}
\examples{
public_trades("binance")
public_trades("binance-us")
public_trades("bitstamp")
public_trades("bittrex")
public_trades("coinbase-pro")
public_trades("crypto.com")
public_trades("ftx")
public_trades("ftx-us")
public_trades("gemini")
public_trades("huobi")
public_trades("kraken")
public_trades("kucoin")
public_trades("poloniex")
}
