% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_celsius_history.R
\name{read_celsius_history}
\alias{read_celsius_history}
\title{Read Celsius History File}
\usage{
read_celsius_history(path)
}
\arguments{
\item{path}{File path to downloaded csv.}
}
\value{

}
\description{
Convenience function to import a csv from Celsius history.
  API access seems difficult at this moment for individuals with Celsius but it is possible
  to export a transaction log manually.
}
