% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key as R Environment Variable}
\usage{
set_api_key(exchange, key_type, key = NULL)
}
\arguments{
\item{exchange}{Which exchange to use for price and market data. Choices are "binance", "binance-us", "bitstamp", "bittrex", "coinbase",
"coinbase-pro", "crypto.com", "ftx", "ftx-us", "gemini", "huobi", "kraken", "kucoin", and "poloniex".}

\item{key_type}{Key type: Either "API", "SECRET", or "PASSPHRASE".}

\item{key}{Leave NULL and input in the askpass popup.}
}
\value{

}
\description{
Set API Key as R Environment Variable
}
