% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_asset_list.R
\name{public_asset_list}
\alias{public_asset_list}
\title{GET asset list for each exchange}
\usage{
public_asset_list(exchange, ...)
}
\arguments{
\item{exchange}{Which exchange to use for list and market data. Choices are "binance", "binance-us", "bitstamp", "bittrex", "coinbase",
"coinbase-pro", "crypto.com", "ftx", "ftx-us", "gemini", "huobi", "kraken", "kucoin", and "poloniex".}

\item{...}{Query parameters passed to API call}
}
\value{

}
\description{
GET asset list for each exchange
}
\examples{
public_asset_list("binance")
public_asset_list("binance-us")
public_asset_list("bitstamp")
public_asset_list("bittrex")
public_asset_list("coinbase")
public_asset_list("coinbase-pro")
public_asset_list("crypto.com")
public_asset_list("ftx")
public_asset_list("ftx-us")
public_asset_list("gemini")
public_asset_list("huobi")
public_asset_list("kraken")
public_asset_list("kucoin")
public_asset_list("poloniex")
}
