% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_ticker_price.R
\name{public_ticker_price}
\alias{public_ticker_price}
\title{GET current public ticker price and market data}
\usage{
public_ticker_price(
  exchange = "binance",
  base_asset = "BTC",
  quote_asset = "USD",
  price_only = TRUE,
  dry_run = FALSE,
  ...
)
}
\arguments{
\item{exchange}{Which exchange to use for price and market data. Choices are "binance", "binance-us", "bitstamp", "bittrex", "coinbase",
"coinbase-pro", "crypto.com", "ftx", "ftx-us", "gemini", "huobi", "kraken", "kucoin", and "poloniex".}

\item{base_asset}{Base asset (default BTC)}

\item{quote_asset}{Quote asset (default USD)}

\item{price_only}{If TRUE, only give current price as a numeric vector. If FALSE, give all market data available from API call.}

\item{dry_run}{If TRUE, call httr2::req_dry_run, which shows the API call without actually sending it.}

\item{...}{Query parameters passed to API call}
}
\value{

}
\description{
GET current public ticker price and market data
}
\examples{
public_ticker_price("binance")
public_ticker_price("binance")
public_ticker_price("binance-us")
public_ticker_price("bitstamp")
public_ticker_price("bittrex")
public_ticker_price("coinbase")
public_ticker_price("coinbase-pro")
public_ticker_price("crypto.com")
public_ticker_price("ftx")
public_ticker_price("ftx-us")
public_ticker_price("gemini")
public_ticker_price("huobi")
public_ticker_price("kraken")
public_ticker_price("kucoin")
public_ticker_price("poloniex")
}
